Rename utilities.* to programmableIOControl.*
