+++ speeduino/init.cpp
@@ -9,7 +9,7 @@
 #include "comms.h"
 #include "comms_secondary.h"
 #include "comms_CAN.h"
-#include "utilities.h"
+#include "programmableIOControl.h"
 #include "scheduler.h"
 #include "schedule_calcs.h"
 #include "auxiliaries.h"
