+++ speeduino/logger.cpp
@@ -3,7 +3,6 @@
 #include "decoders.h"
 #include "init.h"
 #include "maths.h"
-#include "utilities.h"
 #include "preprocessor.h"
 #include "units.h"
 #include "board_definition.h"
