+++ speeduino/sensors.cpp
@@ -18,7 +18,6 @@ A full copy of the license may be found in the projects root directory
 #include "pages.h"
 #include "decoder_init.h"
 #include "auxiliaries.h"
-#include "utilities.h"
 #include "unit_testing.h"
 #include "sensors_map_structs.h"
 #include "units.h"
