+++ speeduino/storage.cpp
@@ -11,7 +11,6 @@ A full copy of the license may be found in the projects root directory
 #include "storage.h"
 #include "pages.h"
 #include "sensors.h"
-#include "utilities.h"
 #include "preprocessor.h"
 #include "unit_testing.h"
