+++ test/test_init/test_fuel_schedule_init.cpp
@@ -3,7 +3,6 @@
 #include "globals.h"
 #include "init.h"
 #include "schedule_calcs.h"
-#include "utilities.h"
 #include "../test_utils.h"
 #include "storage.h"
 #include "../test_schedules/channel_test_helpers.h"
